ml_engine/requirements.txt
flask
flask-cors
pdfminer.six
docx2txt
spacy
scikit-learn
pandas
numpy
python-dotenv
google-generativeai
